requirements.txt
Flask==3.1.3
Werkzeug==3.1.6
Jinja2==3.1.6
MarkupSafe==3.0.3
click==8.3.1
itsdangerous==2.2.0
blinker==1.9.0
pillow==12.1.1
